+++ packages/keychain/src/components/v2/Header/hooks.ts
