+++ packages/keychain/src/components/v2/signup/index.tsx
@@ -1,0 +1,2 @@
+export * from "./Signup";
+export * from "./Login";
