+++ packages/keychain/src/components/v2/signup/validate.ts
@@ -1,0 +1,9 @@
+export function validateUsername(val: string) {
+  if (!val) {
+    return "Username required";
+  } else if (val.length < 3) {
+    return "Username must be at least 3 characters";
+  } else if (val.split(" ").length > 1) {
+    return "Username cannot contain spaces";
+  }
+}
