+++ packages/ui/src/components/index.ts
@@ -1,3 +1,4 @@
 export * from "./icons";
+export * from "./Field";
 
 export * as brand from "./brand";
