+++ packages/ui/src/theme/components/Container.tsx
@@ -2,7 +2,6 @@ import type { ComponentStyleConfig } from "@chakra-ui/theme";
 
 export const Container: ComponentStyleConfig = {
   baseStyle: {
-    py: "16px",
     maxWidth: "container.lg", // 1024px
   },
 };
